Added images for the buttons int ExcelTest

+++ ExcelTest/ExcelTest/Form1.cs
@@ -1,22 +1,18 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.Data;
 using System.Diagnostics;
 using System.Drawing;
 using System.IO;
 using System.Linq;
-using System.Security.Cryptography.X509Certificates;
 using System.Text;
 using System.Threading.Tasks;
 using System.Threading;
 using System.Windows.Forms;
 using Microsoft.Office.Interop.Excel;
 using System.Runtime.InteropServices;
-using Microsoft.Office.Interop.Word;
 using Application = Microsoft.Office.Interop.Excel.Application;
 using Range = Microsoft.Office.Interop.Excel.Range;
-
 namespace ExcelTest
 {
     public partial class Form1 : Form
@@ -24,7 +20,7 @@ namespace ExcelTest
         const string FILE_NOT_FOUND = "Файлът не беше намерен. Проверете дали пътя е верен или файла не е изтрит.";
         const string FILE_NOT_SAVED = "Файлът не беше запаметен.";
         const string FILE_SAVED = "Файлът е променен успешно";
-        const string BLANK_TEXT_BOX = "(Хвани и поостави Excel файла тук)";
+        const string BLANK_TEXT_BOX = "(Хвани и постави Excel файла тук)";
         public Form1()
         {
             InitializeComponent();
@@ -32,28 +28,35 @@ namespace ExcelTest
 
         public void startReplacing_Click(object sender, EventArgs e)
         {
-            
+            if (filePathTextBox.Text == BLANK_TEXT_BOX || filePathTextBox.Text.Trim() == "")
+            {
+                ExceptionLabel.Text = "Преди да стартирате, заредете файла за обработка.";
+                return;
+            }
+
             Cursor = Cursors.WaitCursor;
             ExceptionLabel.Text = "Файлът се отваря....";
             var xlApp = new Application();
+            
             //xlApp = new Microsoft.Office.Interop.Excel.Application();
             try
             {
                 string rememberFilePath = filePathTextBox.Text;
-                var xlWorkBook = xlApp.Workbooks.Open(filePathTextBox.Text, ReadOnly: false, Format: 5,
+                Workbook xlWorkBook = xlApp.Workbooks.Open(filePathTextBox.Text, ReadOnly: false, Format: 5,
                     IgnoreReadOnlyRecommended: true,
-                    Origin: XlPlatform.xlWindows, Delimiter: "\t", Editable: true, Notify: true, AddToMru: true,
+                    Origin: XlPlatform.xlWindows, Delimiter: "\t", Editable: true, Notify: true, AddToMru: false,
                     Local: 1, CorruptLoad: 0);
-                
+
                 Worksheet xlWorkSheet = (Worksheet) xlWorkBook.Worksheets.Item[1];
                 Range range = xlWorkSheet.UsedRange;
                 int rowsRange = range.Rows.Count;
                 int colsRange = range.Columns.Count;
 
-                Thread newThread = new Thread(() => ConvertPZN(rowsRange, range));
-                newThread.Start();
+                //Thread newThread = new Thread(() => );
+                //newThread.Start();
+                ConvertPZN(rowsRange, range);
+                //ExceptionLabel.Text += Environment.NewLine + "Малко форматиране и сме готови....";
 
-                ExceptionLabel.Text = "Малко форматиране и сме готови....";
                 AddBorders(rowsRange, colsRange, range);
                 range.Columns.AutoFit();
                 
@@ -62,32 +65,58 @@ namespace ExcelTest
                     ExceptionLabel.Text = FILE_SAVED;
                 else
                 {
-                    ExceptionLabel.Text = FILE_NOT_SAVED;
+                    ExceptionLabel.Text = FILE_NOT_SAVED +
+                                          Environment.NewLine +
+                                          "Проверете дали файла не се използва от друго приложение.";
                 }
-                xlWorkBook.Close();
+                xlWorkBook.Close(false);
                 xlApp.Quit();
-                newThread.Join();
-                
+                //newThread.Join();
+
                 Cursor = DefaultCursor;
                 Marshal.ReleaseComObject(xlApp);
-                
+
+            }
+            catch (COMException ce)
+            {
+                ExceptionLabel.Text = FILE_NOT_FOUND +
+                                      Environment.NewLine +
+                                      ce.Message;
+                Cursor = DefaultCursor;
             }
             catch (FileNotFoundException exception)
             {
                 ExceptionLabel.Text = FILE_NOT_FOUND +
-                    Environment.NewLine + exception.Message;
+                    Environment.NewLine + 
+                    exception.Message;
                 xlApp.Quit();
+                Cursor = DefaultCursor;
                 Marshal.ReleaseComObject(xlApp);
             }
             catch (Exception exception)
             {
-                ExceptionLabel.Text = exception.Message;
+                //ExceptionLabel.Text = exception.Message;
+                if (exception.Message.Contains("We can't save"))
+                    ExceptionLabel.Text = "Файлът не може да бъде запаметен, защото се използва." +
+                                          Environment.NewLine +
+                                          "Моля затворете файла и опитайте отново";
+                else
+                {
+                    ExceptionLabel.Text = exception.Message;
+                }
+                File.AppendAllText("log.txt", new string('-', 80) +
+                    Environment.NewLine + 
+                    DateTime.Now +
+                    Environment.NewLine +
+                    exception.Message + 
+                    Environment.NewLine + exception.StackTrace);
+
                 Cursor = DefaultCursor;
+                xlApp.Workbooks.Close();
                 xlApp.Quit();
                 Marshal.ReleaseComObject(xlApp);
             }
         }
-
         
         private void ExceptionLabel_Click(object sender, EventArgs e)
         {
@@ -104,7 +133,6 @@ namespace ExcelTest
             string[] files = (string[])e.Data.GetData(DataFormats.FileDrop);
             if (files != null && files.Length != 0)
             {
-                //ExceptionLabel.Text = files[0];
                 filePathTextBox.Text = files[0];
             }
         }
@@ -116,40 +144,38 @@ namespace ExcelTest
 
         private void killExcelButton_Click(object sender, EventArgs e)
         {
-            
             Process[] process = Process.GetProcessesByName("Excel");
-            
-            foreach (var p in process)
-            {                
+            if (MessageBox.Show("Сигурни ли сте, че искате да затворите всички Excel процеси?",
+                "Опасна операция!", MessageBoxButtons.YesNo) == DialogResult.Yes)
+            {
+                foreach (var p in process)
                     p.Kill();
             }
         }
-
         private void filePathTextBox_TextChanged(object sender, EventArgs e)
         {
+            if(filePathTextBox.Text.Trim() == "")
+                filePathTextBox.Text = filePathTextBox.Text.Trim();
             if (filePathTextBox.Text != BLANK_TEXT_BOX)
                 filePathTextBox.ForeColor = Color.Black;
             else
-            {
                 filePathTextBox.ForeColor = Color.LightGray;
-            }
+
             if (filePathTextBox.Text.Contains("pass:"))
             {
+                //filePathTextBox.UseSystemPasswordChar = true;
                 filePathTextBox.PasswordChar = '*';
                 if (filePathTextBox.Text == "pass:adm1n1strat0r")
-                {
                     killExcelButton.Enabled = true;
-                    killExcelButton.Cursor = DefaultCursor;
-                }
+                else
+                    killExcelButton.Enabled = false;
             }
             else
             {
+                filePathTextBox.PasswordChar = '\0';
                 killExcelButton.Enabled = false;
-                killExcelButton.Cursor = Cursors.No;
             }
-
         }
-
         private void filePathTextBox_Enter(object sender, EventArgs e)
         {
             if (filePathTextBox.Text == BLANK_TEXT_BOX)
@@ -157,13 +183,11 @@ namespace ExcelTest
                 filePathTextBox.Text = "";
             }
         }
-
         private void filePathTextBox_Leave(object sender, EventArgs e)
         {
             if (filePathTextBox.Text == "")
                 filePathTextBox.Text = BLANK_TEXT_BOX;
         }
-
         private void loadExcelFile_Click(object sender, EventArgs e)
         {
             openExcelFile.Filter = "Excel Workbook|*.xlsx; *.xlsm|Excel 97-2003 Workbook|*.xls";
